Added support for 2D and ascii versions of Nek files.
Also deprecated the meshcoords tag in the .nek3d file,
getting the info instead while scanning headers for 
time info.

Also added knowledge of misc scalar fields that follow
temperature and pressure.  In the file headers they 
are tagged with 1 2 3 ...  and in Visit they show up
as s1, s2, s3, ...


git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@2495 18c085ea-50e0-402c-830e-de6fd14e8384