Memoize the `MultipassRendering' function.
Turns out that the test for transparencies in actors uses global communication.
Unfortunately IceT reserves the right to call our render function -- or not --
as many times as it wants.  Therefore there is potential for deadlock if we use
global communication during the RenderGeometry or RenderTranslucent calls.
This happens in practice when using HW rendering with IceT.

By memoizing the check for multipass rendering, we avoid subsequent global
communication.


git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@4731 18c085ea-50e0-402c-830e-de6fd14e8384