Removed any asserts which I've let creep in.
Based on visit-developers conversation, we do not want asserts present
in the source tree.  So I did some quick greps and removed any asserts
which were found.  For IceTNM, I also changed the DEBUG_ONLY macro to
utilize an ICET_DEBUG definition instead of NDEBUG.  This reverses the
meaning; a developer must define the macro to add the debugging
support, instead of relying on a macro to remove the debug support.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@5121 18c085ea-50e0-402c-830e-de6fd14e8384