Fix misc bugs for V1.11.
(1) The Threshold operator now issues a warning when applied to the Vector or Truecolor plots (or any other plots that don't take scalars), basically saying that it won't work.  This resolves '8559
(2) I added a check in the avtUnaryMathExpression to guard against de-referencing NULL pointers.  This had a small amount of trickle down to derived types.  This resolves '8621.
(3) I changed the text for the pos_cmfe expression when it gets inserted in the Expressions window.  It now gives the "fill" variable a more meaningful description.  This resolves '8127.
(4) I made the Slice operator label the axes differently when doing an arbitrary slice.  As background, if you do an orthogonal slice by Y=0, it relabels the axes to be "X" and "Z", giving the impression that we are maintaining the axes labels correctly.  And for orthogonal slices, we were.  But for arbitrary slices, we always had "X" and "Y", which created confusion for a user who assumed we were "doing the right thing" based on our track record with orthogonal slices.  I changed the labels to be (Up[0], Up[1], Up[2])-Axis for Y and (Cross[0],Cross[1], Cross[2])-Axis for X (with numbers, not "Up" or "Cross").  The Cross is the cross product of the Up and the Normal.  This resolves '8751.
(5) The same user wanted to use the "theta-phi" slice selection and pointed out that he could not change the up axis.  I enabled this widget.  This resolves '8761.
(6) We were always assuming that 2D Boxlib files were ZR.  I enhanced the reader to read the coordinate system information out of the "Header" file and correctly report the coordinate system.  This resolves '8755.
(7) My machine was renamed, and I changed the config-site files.



git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@5288 18c085ea-50e0-402c-830e-de6fd14e8384