Hacked up version of Tuvok.
This adds a version of Tuvok which has many ad hoc changes in an
effort to get VisIt working as soon as possible.  The API is going
to be redesigned a bit in light of the difficulties, but this
commit at least gets Tuvok available in VisIt to work on in
parallel.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@6604 18c085ea-50e0-402c-830e-de6fd14e8384