Correct an earlier patch I had made to PPC HDF4.
Not entirely sure about this one; I received a patch against
build_visit but it did not apply cleanly, yet I could not discern
why the patch didn't apply.  I eyeballed it and hand-hacked our
patch to follow what the change I was given was.

Thanks again to Patrick Shinpaugh.

(This is a merge from 1.12RC.)


git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@6655 18c085ea-50e0-402c-830e-de6fd14e8384