Do ad hoc memoization of the transparency calculation.
In some SR mode configurations, we have the issue that rendering
will involve global communication due to the need to sort
transparent geometry.  This is undesirable because 1) we don't
know the sort order until rendering is underway, thus preventing
us from using that knowledge for more intelligent compositing, and
2) certain compositors, such as IceT, do not call our render
method synchronously in all cases, leading to deadlock in the
worst case.

This commit changes the transparency actor so that it keeps an
internal cache of the transparency calculation.  Clients are now
expected to clear ("Invalidate") the cache periodically.  In SR
mode we'll do that once per render.  In standard rendering modes,
we'll invalidate the cache every time we add a plot to the window.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@7402 18c085ea-50e0-402c-830e-de6fd14e8384