Add some functions for looking up runtime parameters.
This adds a `RuntimeSettings' namespace which has functions for
easily configuring and looking up runtime parameters.  The
important consideration this implementation handles is finding
parameters from multiple places.

Settings are given a compile time default, and metadata such as
associated environment variables and command line options.  The
search defaults to the compile-time value, but successively looks
for settings which are more specific for the user.

The current ordering, from most important/specific to least, is:
  command line options
  environment variable
  configuration file entry (unimplemented)
  compile-time default

Adding a new parameter is as simple as adding a new entry to a
struct within the implementation.  There is a framework in place
to look up integer, double, boolean, and string values.  String
values are a bit special: if they begin with '.', we prepend
VisIt's installation directory to them.  This allows one to store
settings which point to directories or files.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@8011 18c085ea-50e0-402c-830e-de6fd14e8384