Made gdb at least print out a backtrace.
When not using an xterm for debugging, the gdb we run is not
interactive.  This makes it difficult to figure out where,
exactly, a component crashed.  This adds a backtrace print to our
gdb input file, to make sure we always get at least some useful
output.

Also gdb's startup banner is spammy; -q[uiet] it.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@8013 18c085ea-50e0-402c-830e-de6fd14e8384