Update VTK build to link against OSMesa.
Recent OSMesa contains both mangled symbols plus the standard
OSMesa context management symbols.  Thus we want to route all of
our software-based rendering through OSMesa instead of MesaGL.

This really only works because we never use / don't care about the
X-based software case.  If we have an X server, we always want to
use the native GL library; it's only when we do offscreen
rendering (as we do in SR mode) that we want to force rendering
via Mesa.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@8046 18c085ea-50e0-402c-830e-de6fd14e8384