Bail early and return an error if the dlopen fails.
This allows a client to detect that the dlopen for the library
they gave failed, usually due to a bad path.

The actual error that gets returned is garbage and will change.
Clients should check if glewInitLibrary gives something other than
GLEW_OK instead of testing for this specific error code.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@8727 18c085ea-50e0-402c-830e-de6fd14e8384