Reimplement volume plot on top of GLEW.
Don't do custom #ifdef'ing in an attempt to tease out if Mesa is
in use.  Just check GLEW to see if the extensions work, throw up
error messages if they don't, and we're done.

This allows one to use the volume plot with parallel, hardware
accelerated rendering... though it doesn't matter much, because we
move all the data to processor 0 before rendering.  Still, it
works.

git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@11063 18c085ea-50e0-402c-830e-de6fd14e8384