I corrected the code that adds a "current" symbolic link to a binary distribution created by CPACK, that I then commented out since CPACK doesn't handle symbolic links properly. I am now going to pursue a route of adding the current link as a postprocessing step in the visit-build-open script. When the bug is fixed in cmake the lines can be uncommented. I am not to hopeful since there are several tickets in cmake's bug tracker related to this going back about 3 years.
Information on the CPACK bug:

Currently CPACK removes all symbolic links and replaces files with hard
links. If a symbolic link is to a directory it creates a hard link for
every file in the directory hierarchy that the link points to. Besides
being wasteful it can also create an invalid tar file since it only
stores 100 characters of hard link information in the tar file, which
we run into with some of our files in the include directory.



git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@11601 18c085ea-50e0-402c-830e-de6fd14e8384