Removed setting of SVN_SOURCE_PATH to something other than 'trunk/src'. For a trunk version of build_visit, it should only ever be 'trunk/src.'
Added --disable-dap to netcdf build to eliminate dap and curl dependence.

Fixed patching of netcdf4 to include all 3 places where netcdf sizing
constants need to be changed; libsrc/netcdf.h, libsrc4/netcdf.h and
libsrc4/netcdf_base.h. The only reason I caught this was that another
third party lib, ITAPS-MOAB, also uses netcdf (configured for Exodus)
but also tested the netcdf installation to ensure indeed it had been
patched for ExodusII sizing needs. Those tests were failing and so I
discovered that netcdf did some odd-ball work on the header files 
(I assume just prior to installing) and was installing an UNpatched
header.

Conditioned setting of PAR_INCLUDE_STRING for IceT build for non-Darwin
on whether or not it was already defined. This circumvents the attempt
to run 'mpicxx --shome:compile' which rarely works for the mpich installs
I use.

Note: after all of this, I am still finding a run-time problem with my
ExodusII plugins which I need to investigate and fix. Maybe as simple
as not finding libhdf5.so correctly.


git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@13409 18c085ea-50e0-402c-830e-de6fd14e8384