I adjusted protocol and embedded viewer slightly so they can operate as unit tests.
For protocol, I was getting an Abort on the final line of main on a 
'return 0'. Changing to 'exit(0)' fixed that. I also needed to add
a symlink in bin dir for visitprotocol.

For embedded viewer, I needed to adjust it to take name of command file
on command line and then execute those commands at startup.

I then added tests for visitprotocol and embedded viewer.

We may find the protocol test results in perhaps more re-baselining than
is convenient as it involves most (if not all) of the state objects we
ship around. We could maybe fix this by some filtering out of 'unstable'
parts of VisIt and/or not baselining the test's output but instead just
its return code.

We may find the embedded viewer test cannot run from non-interactive logins.
I don't know. It brings up a window and does its work there. I don't know
if that will NOT work for nightly testing.


git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@13487 18c085ea-50e0-402c-830e-de6fd14e8384