ordered compositing
add alpha blend compositing to visit and use it to implement
ordered compositing. there is some logic to enable it automatically
and a gui option to explicitly disable it. it is on by default.
I also added an option for enabling depth peeling. this is off by
default because it's very slow. I changed all GetCell methods to
return long long so that overflow is less of an issue. I fixed a bug
in the transparency actor that was causing transparent geometry
to be sorted twice.


git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@27514 18c085ea-50e0-402c-830e-de6fd14e8384